Update: move marked functions into core
