+++ test/mutation/data/one_res_P1G.pdb
@@ -1,0 +1,6 @@
+ATOM      1  N   GLY A   1       4.705  11.003  -7.826  1.00 27.02           N  
+ATOM      2  CA  GLY A   1       3.861  11.697  -6.851  1.00 25.84           C  
+ATOM      3  C   GLY A   1       4.023  11.142  -5.443  1.00 25.58           C  
+ATOM      4  O   GLY A   1       4.668  10.102  -5.236  1.00 24.12           O  
+TER
+END
